todos los valores para ocupar (sin validación)
