+++ src/app/componentes/aseguradoras/aseg-datos-emision/datos-emision-cliente/datos-emision-cliente.component.html
@@ -1,0 +1,79 @@
+<div class="card mb-4 rounded" style="justify-content: center">
+  <div class="card-body">
+    <div class="container">
+      <div class="row">
+        <div class="col-12">Cliente</div>
+      </div>
+      <div class="row">
+        <div class="form-group col-4">
+          <label class="form-label">Nombre:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Apellido Paterno:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Apellido Materno:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">RFC:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Correo:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Ocupación:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Fecha de Nacimiento:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Genero:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Teléfono:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Nacionalidad:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Código Postal:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Estado:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Municipio:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Colonia:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Calle:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Numero Exterior:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Numero Interior:</label>
+          <input type="text" class="form-control" />
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
