+++ src/app/componentes/aseguradoras/aseg-datos-emision/datos-emision-cliente/datos-emision-cliente.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-datos-emision-cliente',
+  templateUrl: './datos-emision-cliente.component.html',
+  styleUrls: ['./datos-emision-cliente.component.css']
+})
+export class DatosEmisionClienteComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
