+++ src/app/componentes/aseguradoras/aseg-datos-emision/datos-emision-vehiculo/datos-emision-vehiculo.component.html
@@ -1,0 +1,39 @@
+<div class="card mb-4 rounded" style="justify-content: center">
+  <div class="card-body">
+    <div class="container">
+      <div class="row">
+        <div class="col-12">Vehículo</div>
+      </div>
+      <div class="row">
+        <div class="form-group col-4">
+          <label class="form-label">Tipo:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Año:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Marca :</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Descripción:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Número de Serie:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Número de Placa:</label>
+          <input type="text" class="form-control" />
+        </div>
+        <div class="form-group col-4">
+          <label class="form-label">Número de Motor:</label>
+          <input type="text" class="form-control" />
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
