+++ src/app/componentes/aseguradoras/aseg-datos-emision/datos-emision-vehiculo/datos-emision-vehiculo.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-datos-emision-vehiculo',
+  templateUrl: './datos-emision-vehiculo.component.html',
+  styleUrls: ['./datos-emision-vehiculo.component.css']
+})
+export class DatosEmisionVehiculoComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
