+++ src/app/componentes/pagina3/pagina3.component.html
@@ -1,3 +1,4 @@
+<!--
 <div class="container mw-90">
   <app-infovehiculo
     [disabled]="noedites"
@@ -113,9 +114,16 @@
   <app-ase-anna *ngIf="nombraseguradora == 'ANA SEG'"></app-ase-anna>
   <app-ase-aig *ngIf="nombraseguradora == 'AIG MEX'"></app-ase-aig>
 </div>
+-->
 
 <div class="container mw-90">
   <app-aseg-datos-emision
     [aseguradora]="nombraseguradora"
   ></app-aseg-datos-emision>
+</div>
+
+<div class="row justify-content-center">
+  <button class="btn btn-primary btn-lg" onclick="alert('Emitiendo')">
+    EMITIR
+  </button>
 </div>
